+++ include/material.h
@@ -9,7 +9,7 @@ private:
 	Pixel attenuation;
 public:
 	Material(const Pixel & attenuation);
-	virtual Vec3 scatter(const Ray & r, const gObject * obj, const double t) const = 0;
+	virtual Vec3 scatter(const Ray & r, const gObject * obj, const float t) const = 0;
 	Pixel getAttenuation();
 };
