+++ src/pixel.cc
@@ -3,7 +3,7 @@
 using namespace std;
 
 ostream & operator << (ostream & out, const Pixel & pxl) {
-  double c = 255.99;
+  float c = 255.99;
   out << int(pxl.r_ * c) << ' ' << int(pxl.g_ * c) << ' ' << int(pxl.b_ * c) << endl;
   return out;
 }
@@ -16,13 +16,13 @@ Pixel operator*(const Pixel & p1, const Pixel & p2) {
   return Pixel(p1.r() * p2.r(), p1.g() * p2.g(), p1.b() * p2.b());
 }
 
-Pixel operator/(const Pixel & p1, double c) {
+Pixel operator/(const Pixel & p1, float c) {
   return Pixel(p1.r() / c, p1.g() / c, p1.b() / c);
 }
-Pixel operator*(const Pixel & p1, double c) {
+Pixel operator*(const Pixel & p1, float c) {
   return Pixel(p1.r() * c, p1.g() * c, p1.b() * c);
 }
-Pixel operator*(double c, const Pixel & p1) {
+Pixel operator*(float c, const Pixel & p1) {
   return Pixel(p1.r() * c, p1.g() * c, p1.b() * c);
 }
 
@@ -32,7 +32,7 @@ void Pixel::operator+=(const Pixel & p2) {
   b_ += p2.b_;
 }
 
-void Pixel::operator/=(double c) {
+void Pixel::operator/=(float c) {
   r_ /= c;
   g_ /= c;
   b_ /= c;
